15/02/2020

+++ petClinic/src/app/components/pets/pets.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit, Input } from "@angular/core";
 import { Pet } from "../../models/pet";
-import { Owner } from "src/app/models/owner";
 import { PetService } from "src/app/services/pet.service";
 import { Router } from "@angular/router";
 
@@ -10,20 +9,22 @@ import { Router } from "@angular/router";
   styleUrls: ["./pets.component.css"]
 })
 export class PetsComponent implements OnInit {
-  @Input() owner: Owner;
-  private pets: Pet[];
+  @Input() pet: Pet;
   constructor(private petService: PetService, private route: Router) {}
 
   deletePet(pet: Pet) {
+    if (pet.visits.length > 0) {
+      alert(
+        "It is not possible to remove pets with visits. Delete visits before"
+      );
+      return false;
+    }
     if (confirm(" Do you want to delete a " + pet.name + " ?")) {
       this.petService.deletePet(pet.id).subscribe();
-      this.route.navigate(["/owners/" + this.owner.id]);
     }
   }
 
   ngOnInit() {
-    this.petService.getPetsOwnerId(this.owner.id).subscribe(respuesta => {
-      this.pets = respuesta;
-    });
+    console.log(this.pet);
   }
 }
